Edit, update e delete arrumados

+++ app/views.py
@@ -143,12 +143,17 @@ def edit(request, pk):
 # Função para atualizar edições de atestados
 def update(request, pk):
     if request.user.is_authenticated:
-        data = {}
-        data['db'] = Atestados.objects.get(pk=pk)
-        form = AtestadosForm(request.POST, request.FILES or None, instance=data['db'])
-        if form.is_valid():
-            form.save()
-            return redirect('home')
+        dado = Atestados.objects.get(pk=pk)
+        if dado.user == request.user:
+            data = {}
+            data['db'] = Atestados.objects.get(pk=pk)
+            form = AtestadosForm(request.POST, request.FILES or None, instance=data['db'])
+            if form.is_valid():
+                form.save()
+                return redirect('home')
+        else:
+            messages.error(request, 'O atestado selecionado não pertence ao usuário atual, portanto, este não está autenticado para realizar a ação!')
+            return redirect('pesquisa')
     else:
         messages.error(request, 'Usuário não conectado!')
         return redirect('entrar')
@@ -156,9 +161,14 @@ def update(request, pk):
 # Função para deletar atestados
 def delete(request, pk):
     if request.user.is_authenticated:
-        db = Atestados.objects.get(pk=pk)
-        db.delete()
-        return redirect('pesquisa')
+        dado = Atestados.objects.get(pk=pk)
+        if dado.user == request.user:
+            db = Atestados.objects.get(pk=pk)
+            db.delete()
+            return redirect('pesquisa')
+        else:
+            messages.error(request, 'O atestado selecionado não pertence ao usuário atual, portanto, este não está autenticado para realizar a ação!')
+            return redirect('pesquisa')
     else:
         messages.error(request, 'Usuário não conectado!')
         return redirect('entrar')
